+++ extension/language-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @pandacss/language-server
 
+## 0.11.1
+
 ## 0.11.0
 
 ## 0.10.0