+++ packages/astro/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pandacss/astro
 
+## 0.11.1
+
+### Patch Changes
+
+- @pandacss/postcss@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes