+++ packages/fixture/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pandacss/fixture
 
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies [23b516f4]
+  - @pandacss/types@0.11.1
+  - @pandacss/preset-base@0.11.1
+  - @pandacss/preset-panda@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes